sigtran: Add API osmo_ss7_as_select_asp()
This API is required by osmo-bsc, since when in SCCPLite configuration,
it needs to access the IPA conn of the AS towards the MSC in order to
inject MGCP and CTRL protocols over the IPA multiplex.

This API may need to be revisited (i.e. deprecated and a new one
created) once we want to fully support loadshare, where message contents
such as OPC,DPC,SLS need to be taken into account internally in
libosmo-sigtran.
That new version may well be kept private though, since the API here
presented is good enough for osmo-bsc IPA multiplex injection.
This is anyway kind of chicken-and-egg problem, because we want to have
an API osmo-bsc can use in first place before moving the whole
osmo_ss7_as struct as private so that we can modify it to implement
proper AS loadsharing.

Related: SYS#6602
Change-Id: I8b866b45ac4a24238c55171d25e11b9625e8f50c