AS loadsharing: Introduce combined_linkset
A "Combined linkset" is a set of routes for different AS or linksets
whose destination (DPC & mask) and priority are the same.

An SP is expected to load share traffic among linksets/AS of a given
combined linkset, based on DPC/OPC,SLS. This means certain state or
heuristics need to be stored in the combined linkset in order to keep
forwarding messages of the same transaction (SLS) over the same specific
AS/linkset.

This commit introduces the osmo_ss7_combined_linkset object, which
is roughly a set of osmo_ss7_route sharing DPC+MASK+PRIO key, and
it is placed in between the osmo_ss7_route_table and osmo_ss7_route
entries.
User still operates/manages indiviual routes, but those are grouped
internally inside its corresponding combined linkset.

No routing change is implemented in this commit. Code is only tweaked
minimally so far to retrieve the first route in the combined link.
This will allow future work, i.e to:
- Check if data can be sent over the combined link (at least one
  AS/linkset is ACTIVE), and otherwise try using a combined link with a
  less specific match (shorter prefix mask or/and lower priority).
- Select one of the AS/linksets inserted the combined link, based on
  OPC/DPC/SLS of the message.

Related: SYS#7112
Change-Id: I500ec74ba975e3c93071771027e4e5fe6000e6f3