AS loadsharing: Initial routing implementation based on extended-SLS
This commit adds a new set of APIs and structs to support routing based
on OPC,DPC,SLS fields and hence to support AS-loadshare.

The routing decision is still actually left as beforehand, ie. the first
AS/linkset in the combined_linkset is selected.

This will allow future work, i.e to:
    - Check if data can be sent over the combined link (at least one
      AS/linkset is ACTIVE), and otherwise try using a combined link with a
      less specific match (shorter prefix mask or/and lower priority).
    - Select one of the AS/linksets in the combined link based on a
      round robin approach.

Related: SYS#7112
Change-Id: I0fb4ca4959096f748a23082efa0481300de56436