default_lm_fsm: Trigger ASP restart by disconnecting the stream
Previous behavior was to restart the ASP top-to-bottom in the stack,
which doesn't play well with the disconnect_cb().
For instance, upon LM FSM timer callback, osmo_ss7_asp_restart() would call
osmo_stream_cli_open() which in turn would call osmo_stream_close(). As
a result, in that code path events were being sent to the LM FSM itself
notifying the disconnection of the SCTP conn, which in turn would call
osmo_ss7_asp_restart() again, and everything ended up in a weird state
where it couldn't create sockets properly.

The better approach is to, from LM FSM, trigger disconnection of the
SCTP conn, which will in turn, through disconnect_cb(), send the SCTP
disconnect notification up the stack, which should then trigger
re-establishment of the SCTP conn in the SCTP client case.

Change-Id: I21ebc3cd716a0bcc74da0a789032570f038e119d