Fix osmo_ss7_tmode_to_xua(OSMO_SS7_AS_TMOD_ROUNDROBIN)
As specified in RFC4666, "roundrobin" is not a M3UA traffic mode but
simply a specific possible implementation of "loadshare"
traffic mode.
Hence, when converting to M3UA Traffic Mode Type, if AS was configured
by user as "roundrobin", on the wire it is actually configured as
"loadshare" traffic mode.

Change-Id: I56447024936c8518cfd3f947971e3d7d0616e4f7