ipa: Fix asp_id/SLS unset in tcp-server mode
Commit changing the value set to asp_id (used to derive SLS) for IPA
conns only updated the asp_id value on the stream_cli code path, but not
on the stream_srv one.
As a result, on IPA conns with an underlaying TCP server conn, asp_id
and SLS was always 0.

Fixes: 2c9ba16 "ipa: Use pseudo-random number for SLS in IPA->M3UA direction
Change-Id: I6e166aa67da6103d541084a45e8dc5595241fe60