sccp: Allow marking UNIT-DATA.req Sequence Control param non-presence, set SLS
The public API for SCCP UNIT-DATA.req was missing a way to state whether
the "Sequence Control" Parameter was present or not in the primitive.

This was originated most probably by the fact that the field is always
present for SUA CLDT message generated and sent over the wire.

Still, the presence of such field during UNIT-DATA.req actually
indicates the Protocol Class to be used/set on the message.
Hence, it is important to provide it.

Once the Protocol Class to transmit the message is known, we can then
generate a proper SLS.

Change-Id: I834e91c1ec337713b0e684cf4fd7de854150bef6