ASP loadsharing: Normal ASP distribution regardless of active state
Before this patch, the normal ASP for a given loadshare eSLS entry (seed)
would only be filled in with active ASPs.
That means that if eg. N ASPs are configured but only 1 is successfully
activated upon startup and a lot of traffic happens, the full eSLS will
be filled with that ASP, and as a result all traffic will be sent to
that ASP, with the exception during the time that ASP is inactive.

Instead, we want to populate the eSLS table with Normal ASP field
properly distributed over all ASPs. If a given ASP is not active at that
time, then simply pick an Alternative ASP during that time.

Change-Id: I2064fe89bac2d3bd86fee544a64ab21417e1c368