AS loadsharing: Normal route distribution regardless of active state
Before this patch, the normal route for a given loadshare eSLS entry (seed)
would only be filled in with available routes.
That means that if eg. N routes are configured but only 1 is really
available upon startup and a lot of traffic happens, the full eSLS will
be filled with that route, and as a result all traffic on that combined
linkset will be sent through that route.

Instead, we want to populate the eSLS table with Normal Route field
properly distributed over all routes. If a given route is not active at
that time, then simply pick an Alternative Route during that time.

Change-Id: I6daf114416c69c84bdc3042efcaf2918f91c2e87