asp: Rename and clarify asp_id field
The asp_id (RFC4666 ASP Identifier) can be (and may be) different on
each SGP peer of an ASP. Previously, we were storing in the same field
the value provided by the peer, and using this same value when sending
our own ASP Identifier during ASPUP.
This was not causing a problem since actually we have no real way to set
asp_id_present on our own in ASP mode yet, so no local ASP Identifier is
ever being transmitted.

This patch renames the asp_id field to remote_asp_id, since that's the
meaning it has in almost all places where it is being used.
The only place where it is being used so far, during Tx ASPUP, has been
documented and disabled since in reality the asp_id_present was not set
as mentioned above.

Furthermore, the remote ASP Identifier we receive from the peer ASP is
actually meaningful within a given AS, meaning that if an ASP is part of
several AS, we should be storing 1 remote ASP identifier for each AS it
serves. This is documented in ss7_asp7.h for later

Change-Id: I8e4304099db6fb12cf4b40a5b9a4aa3a83c95fde