ipa: Store ASP IPA SLS into its own field
Storing the 4-bit SLS into asp_id field is a total hack which just
pollutes the code and makes it more difficult to understand.

Add a new field into ss7_asp to store that information cleanly.

Moreover, asp_id field itself is right now a total mess on its own,
since it is used both for local and remote ASP Identifiers at the same
time. This will be fixed in a follow-up commitset.

Change-Id: I6ec346417ceeabfce9c7b856040be2179c59195e