asp: Make sure previous stream_srv is immediatelly destroyed when requested
This will be needed when the asp is in SCTP=server mode and we request
it to shutdown/restart.

Path in xua_accept_cb() for dynamic ASPs in SCTP=server mode needs to be
fixed since it was wrongly calling osmo_ss7_asp_restart() and taking
advantage of the fact that the asp->server was not being freed in that
case.

Change-Id: I8edd64234654ba987f35de2d7ad610e96bda27eb