xua_snm.c: Handle DAUD with aff_pc containing wildcards
Handle it with a really basic implementation for completeness.
In theory we shouldn't in general be receiving such big masks,
since RFC466 3.4.3 states:
"""
It is recommended that during normal operation (traffic handling) the
mask field of the Affected Point Code parameter in the DAUD message
be kept to a zero value in order to avoid SG overloading.
"""

Change-Id: Icdb726ae66d3bee19b2113b1e77d68b517e8fb4d