sccp: Split sccp_scoc into multiple files
The previous state was really messy with sccp_scoc.c containing lots
of different not-directly-related modules, which made it very
difficult to find all the parts and gasp the whole picture.
It also makes it easier to extend the code in the future as we plan to
be adding more features from time to time.

This commit moves (splits) sccp_scoc.c + sccp_internal.h out to
different modules:
* sccp_user.h is added since there already existed sccp_user.c,
  operating on struct sccp_user.
* sccp_connection.{c,h} is added to contain all the struct
  sccp_connection data domain logic.
* sccp_scoc_fsm.{c,h} is added to have all the related FSM code
  self-contained in a file.
* sccp_sap.c: Some code related to the SCU primitive is moved here,
  where it belongs.
* sccp_vty: Some related vty code is moved here, where it belongs.

The SCOC handling logic is kept in the sccp_scoc.c file.
Some functions are properly prefixed when moved to the correct file.

Change-Id: I51e9bf60bc4c4b2babd3f408daed914eaf398e82