sccp_sclc: Remove SCCP cause mask in N-NOTICE.ind 'Reason for return' field
ITU Q.713 A.5 states "There is a one-to-one mapping between the return cause of
UDTS, XUDTS or LUDTS messages and the reason for return in the N-NOTICE
primitives".
This matches with the list of strings listed in ITU Q.711 6.2.2.2.4
"Reason for return".
"return cause" can in turn be found in ITU Q.713 3.12.

During sccp_sclc_rx_scrc_rout_fail(), we are encoding the "return cause"
param into the xua_msg usig gen_ret_msg(), which encodes it as SUA by
prepending the SUA_CAUSE_T_RETURN bitmask cause-type to it.
However, when constructing the primitive N-NOTICE.ind, we need to remove
that mask when filling in "Reason for return".

Change-Id: Iccff1e8809769edae36081a79119eb8da901841a