mtp: Support MTP-TRANSFER.req/ind of raw IPA messages
This is needed ie. in SCCPLite, where MGCP and CTRL messages need to be
sent/received over the same IPA multiplex TCP conn as the SCCPLite
(SCCP/IPA) one towards a given MSC.

Prior to this, osmo-bsc-sccplite was doing lots of manual
libosmo-sigtran APIs to obtain a route, an AS, and ASP, sending over an
ASP, etc. Similary, the rx_unknown callback was added to libosmo-sigtran
to hook into the ASP level. Tons of stuff which should really be done
properly inside libosmo-sigtran and provided over the existing public
SAPs, as this patch proposes.

With the proposed approach, an SCCPLite client (eg. osmo-bsc), besides
setting up an sccp_user to handle SCCP, also sets an ss7_user (MTP SAP)
with the newly introduced new Service Indicator (SI) MTP_SI_NI11_OSMO_IPA
to be able to manage the IPA[MGCP/CTRL] traffic towards a given MTP peer
(SCCPLite client, ie. MSC).

The new Service Indicator (SI) MTP_SI_NI11_OSMO_IPA is added with a
value of 0x02 (reserved in specs) to avoid collision with existing
protocols even if NI=International is used.
Ideally, from conceptual point of view when using SCCPLite one would
configure an ss7 instance with NI=National, or even better
NI=Spare-National, since that SI doesn't exist in International networks.
In practice, it doesn't matter much since this SI is the end only used
in IPA ASP, which are not really standarized, and the SI is not even
transmitted through the wire with IPA, hence it's only a matter of
local configuration.

At an SGP (STP), the changes included in this commit also allow now
routing of IPA frames (other than SCCP), which are routed based on
ASP/AS routing configuration. This means now one could envision having
an SCCPlite BSC connected to an SCCPlite MSC using an STP in between.

Related: SYS#6880
Change-Id: I9fedb26ccd3434fc7f272feb3c45cf4bdb80c7ae