mtp: Implement already defined osmo_mtp_prim_name()
Implement osmo_mtp_prim_name(), which was defined in a header
file but didn't exist so far.

Create a mtp_sap.c file as a counterpart of mtp_sap.h header, and move
there all primitive/SAP related code, to make it easier to spot where
the inter-layer communication happens. This way we also have the same
disposition as with sccp_sap/sccp_user.

This is a preparation commit, follow-up commits will use the
osmo_mtp_prim_name() API and clean up and move more APIs to mtp_sap.c.

Change-Id: Ifa21c26ceaa079435a2d157f99813e06cfc12e71