ipa: Adapt handling logic of rx msg from osmo_ipa_segmentation_cb()
We migrated to using osmo_io+osmo_stream with
osmo_ipa_segmentation_cb(), which takes care of pulling the multiple IPA
headers and stores them in osmo_ipa_msgb_cb_proto(_ext)().
When doing the change, we forgot to update the rx path in ipa.c to
account for those changes. Do it now to simplify the code and clarify
that l1h was actually pulled from ->data.

Change-Id: I4e26bf471de9a258dc2f862488253f31aa95a013